Fix lines
